requirements.txt
# Core Dependencies
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.2
python-multipart>=0.0.6

# AI and Vector Processing
torch>=2.1.1
transformers>=4.35.2
numpy>=1.24.3
pillow>=10.1.0

# Database
neo4j>=5.14.1

# Testing and Development
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2
python-dotenv>=1.0.0

# Utilities
loguru>=0.7.2
tenacity>=8.2.3 